requirements.txt
flask
pandas
torch
transformer